Move _get_module_data() to common.py
