+++ plugwise/common.py
@@ -4,6 +4,7 @@ Plugwise Smile protocol helpers.
 """
 from __future__ import annotations
 
+from plugwise.constants import ModelData
 from plugwise.util import check_heater_central, check_model
 
 from defusedxml import ElementTree as etree
@@ -16,6 +17,7 @@ class SmileCommon:
     def __init__(self) -> None:
         """Init."""
         self._appliances: etree
+        self._cooling_present: bool
         self._heater_id: str
         self._on_off_device: bool
         self._opentherm_device: bool
@@ -25,12 +27,11 @@ class SmileCommon:
         """Helper-function checking the smile-name."""
         return self.smile_name == name
 
-    def _appl_thermostat_info(self, xml: etree, appl: Munch) -> Munch:
+    def _appl_thermostat_info(self, xml_1: etree, xml_2: etree, appl: Munch) -> Munch:
         """Helper-function for _appliance_info_finder()."""
-        # Collect thermostat device info
         locator = "./logs/point_log[type='thermostat']/thermostat"
         mod_type = "thermostat"
-        module_data = self._get_module_data(xml, locator, mod_type)
+        module_data = self._get_module_data(xml_1, xml_2, locator, mod_type)
         appl.vendor_name = module_data["vendor_name"]
         appl.model = check_model(module_data["vendor_model"], appl.vendor_name)
         appl.hardware = module_data["hardware_version"]
@@ -39,13 +40,20 @@ class SmileCommon:
 
         return appl
 
-    def _appl_heater_central_info(self, xml_1: etree, xml_2: etree, appl: Munch) -> Munch:
+    def _appl_heater_central_info(
+        self,
+        xml_1: etree,
+        xml_2: etree,
+        xml_3: etree,
+        appl: Munch
+    ) -> Munch:
         """Helper-function for _appliance_info_finder()."""
         # Remove heater_central when no active device present
         if not self._opentherm_device and not self._on_off_device:
             return None
 
         # Find the valid heater_central
+        # xml_1 = self._appliances for legacy, self._domain_objects for actual
         self._heater_id = check_heater_central(xml_1)
 
         #  Info for On-Off device
@@ -57,16 +65,70 @@ class SmileCommon:
 
         # Info for OpenTherm device
         appl.name = "OpenTherm"
-        locator1 = "./logs/point_log[type='flame_state']/boiler_state"
-        locator2 = "./services/boiler_state"
+        locator_1 = "./logs/point_log[type='flame_state']/boiler_state"
+        locator_2 = "./services/boiler_state"
         mod_type = "boiler_state"
-        module_data = self._get_module_data(xml_2, locator1, mod_type)
+        # xml_2 = appliance
+        # xml_3 = self._modules for legacy, self._domain_objects for actual
+        module_data = self._get_module_data(xml_2, xml_3, locator_1, mod_type)
         if not module_data["contents"]:
-            module_data = self._get_module_data(xml_2, locator2, mod_type)
+            module_data = self._get_module_data(xml_2, xml_3, locator_2, mod_type)
         appl.vendor_name = module_data["vendor_name"]
         appl.hardware = module_data["hardware_version"]
         appl.model = module_data["vendor_model"]
         if appl.model is None:
-            appl.model = "Generic heater"
+            appl.model = (
+                "Generic heater/cooler"
+                if self._cooling_present
+                else "Generic heater"
+            )
 
         return appl
+
+    def _get_module_data(
+        self, xml_1: etree, xml_2: etree, locator: str, mod_type: str, legacy=False,
+    ) -> ModelData:
+        """Helper-function for _energy_device_info_finder() and _appliance_info_finder().
+
+        Collect requested info from MODULES.
+        """
+        model_data: ModelData = {
+            "contents": False,
+            "firmware_version": None,
+            "hardware_version": None,
+            "reachable": None,
+            "vendor_name": None,
+            "vendor_model": None,
+            "zigbee_mac_address": None,
+        }
+        # xml_1 = appliance
+        if (appl_search := xml_1.find(locator)) is not None:
+            link_id = appl_search.attrib["id"]
+            loc = f".//services/{mod_type}[@id='{link_id}']...."
+            if legacy:
+                loc = f".//{mod_type}[@id='{link_id}']...."
+            # Not possible to walrus for some reason...
+            # xml_2 = self._modules for legacy, self._domain_objects for actual
+            module = xml_2.find(loc)
+            if module is not None:  # pylint: disable=consider-using-assignment-expr
+                model_data["contents"] = True
+                if (vendor_name := module.find("vendor_name").text) is not None:
+                    model_data["vendor_name"] = vendor_name
+                    if "Plugwise" in vendor_name:
+                        model_data["vendor_name"] = vendor_name.split(" ", 1)[0]
+                model_data["vendor_model"] = module.find("vendor_model").text
+                model_data["hardware_version"] = module.find("hardware_version").text
+                model_data["firmware_version"] = module.find("firmware_version").text
+                if legacy:
+                    # Stretches
+                    if (router := module.find("./protocols/network_router")) is not None:
+                        model_data["zigbee_mac_address"] = router.find("mac_address").text
+                    # Also look for the Circle+/Stealth M+
+                    if (coord := module.find("./protocols/network_coordinator")) is not None:
+                        model_data["zigbee_mac_address"] = coord.find("mac_address").text
+                # Adam
+                elif (zb_node := module.find("./protocols/zig_bee_node")) is not None:
+                        model_data["zigbee_mac_address"] = zb_node.find("mac_address").text
+                        model_data["reachable"] = zb_node.find("reachable").text == "true"
+
+        return model_data
